new component for model dashboard
