+++ ui/litellm-dashboard/src/components/add_model/litellm_model_name.tsx
@@ -5,9 +5,9 @@ import { Row, Col } from "antd";
 import { Providers } from "../provider_info_helpers";
 
 interface LiteLLMModelNameFieldProps {
-  selectedProvider: string;
+  selectedProvider: Providers;
   providerModels: string[];
-  getPlaceholder: (provider: string) => string;
+  getPlaceholder: (provider: Providers) => string;
 }
 
 const LiteLLMModelNameField: React.FC<LiteLLMModelNameFieldProps> = ({
@@ -39,7 +39,7 @@ const LiteLLMModelNameField: React.FC<LiteLLMModelNameFieldProps> = ({
           {(selectedProvider === Providers.Azure) || 
            (selectedProvider === Providers.OpenAI_Compatible) || 
            (selectedProvider === Providers.Ollama) ? (
-            <TextInput placeholder={getPlaceholder(selectedProvider.toString())} />
+            <TextInput placeholder={getPlaceholder(selectedProvider)} />
           ) : providerModels.length > 0 ? (
             <AntSelect
               allowClear
@@ -67,7 +67,7 @@ const LiteLLMModelNameField: React.FC<LiteLLMModelNameFieldProps> = ({
               style={{ width: '100%' }}
             />
           ) : (
-            <TextInput placeholder={getPlaceholder(selectedProvider.toString())} />
+            <TextInput placeholder={getPlaceholder(selectedProvider)} />
           )}
         </Form.Item>
